revert: cancel stale Equihash work promptly
Reverts zakura-core/zakura#643 so the change can return to draft review.
