--- docs/changelog/unreleased/643.md
@@ -1,5 +1,0 @@
-## Fixed
-
-- Stopped the internal miner from submitting solutions for stale templates and
-  reduced stale work by cancelling between Equihash digit rounds
-  ([#643](https://github.com/zakura-core/zakura/pull/643)).
